// data/github/current/ui/packages/issue-metadata/components/labels/__tests__/__generated__/LabelsSectionTestQuery.graphql.ts
export type LabelsSectionTestQuery$variables = {
  number: number;
  owner: string;
  repo: string;
};
export type LabelsSectionTestQuery$data = {
  readonly repository: {
    readonly issue: {
      readonly " $fragmentSpreads": FragmentRefs<"LabelsSectionFragment">;
    } | null | undefined;
  } | null | undefined;
};
export type LabelsSectionTestQuery = {
  response: LabelsSectionTestQuery$data;
  variables: LabelsSectionTestQuery$variables;
};